Added basic pages.
