+++ saylua/modules/general/templates/pages/about.html
@@ -1,0 +1,9 @@
+{% extends "layout.html" %}
+{% set page_title = "About Saylua" %}
+
+{% block body %}
+<h1>About Saylua</h1>
+<p>
+  Under construction.
+</p>
+{% endblock %}
