+++ saylua/modules/general/templates/pages/contact.html
@@ -1,0 +1,17 @@
+{% extends "layout.html" %}
+{% set page_title = "Contact Us" %}
+
+{% block body %}
+<h1>Contact Us</h1>
+<p>
+  Staff on Saylua can generally be contacted through the site itself. Make sure
+  to contact the proper the staff member for your inquiry.
+</p>
+
+<h2>User Issues</h2>
+<p>
+  If you have a problem with another user on Saylua or see something that you
+  feel breaks the site's rules, please contact a moderator.
+</p>
+
+{% endblock %}
