+++ saylua/modules/general/templates/pages/credits.html
@@ -1,0 +1,35 @@
+{% extends "layout.html" %}
+{% set page_title = "Saylua Credits" %}
+
+{% block body %}
+<h1>Saylua Credits</h1>
+<p>
+  Saylua was primarily created by Mike and Tiff though the project would not
+  have been possible without the contributions of many individuals.
+</p>
+<p>
+  Most of the art for Saylua, including the pet art, was wonderfully drawn by
+  <a href="http://www.chutkat.com/">Chutkat</a> who also designed the overall
+  visual style for the site.
+</p>
+<p>
+  The world of Saylua was brought to life through the beautiful environment
+  illustrations of <a href="http://lyraina.com">Lyraina</a>.
+</p>
+<p>
+  Many additional artists including <a href="http://fyne.deviantart.com/">Fyne</a>,
+  <a href="https://synthaphonograph.tumblr.com/">Synthaphonograph</a>,
+  <a href="https://cheetahtrout.tumblr.com/">Cheetahtrout</a>,
+  and <a href="http://fasius.deviantart.com/">Fasius</a> contributed many
+  important pieces of art early on in the project's life.
+</p>
+<p>
+  Much of Saylua's lore was expanded upon through writing contributions from
+  <a href="http://theartsyaxolotl.com/">ArtsyAxolotl</a>.
+</p>
+<p>
+  And a special mention to the early members of the project who helped
+  significantly shape the direction of Saylua with their feedback, ideas, and
+  encouragement: Namine, Gam, Tears, Inu, Vik, Jamew, and Queenie.
+</p>
+{% endblock %}
