+++ saylua/modules/general/templates/pages/terms.html
@@ -1,0 +1,9 @@
+{% extends "layout.html" %}
+{% set page_title = "The Terms of Service" %}
+
+{% block body %}
+<h1>Terms of Service for Saylua</h1>
+<p>
+  Under construction.
+</p>
+{% endblock %}
