+++ saylua/modules/general/views/__init__.py
@@ -1,0 +1,1 @@
+import main, news # noqa
