+++ saylua/templates/layout.html
@@ -105,11 +105,10 @@
     <p>
       <a href="/online/">{{users_online_count|pluralize('User')}} Online</a> &bull;
       <a href="#">Back to Top</a> &bull;
-      <a href="#">About</a> &bull;
-      <a href="#">Credits</a> &bull;
-      <a href="#">Terms of Service</a> &bull;
-      <a href="#">Contact Us</a> &bull;
-      <a href="#">Staff</a> &bull;
+      <a href="/page/about/">About</a> &bull;
+      <a href="/page/credits/">Credits</a> &bull;
+      <a href="/page/terms/">Terms of Service</a> &bull;
+      <a href="/page/contact/">Contact Us</a> &bull;
       <span id="social-icons">
         <a href="https://www.facebook.com/officialsaylua/" target="_blank">
           <i class="fa fa-fw fa-facebook" aria-hidden="true"></i>
